날짜 변수 context로 관리하도록 수정
