+++ docs/src/contexts/date_context.js
@@ -1,0 +1,5 @@
+
+import { createContext } from "react";
+
+export const MonthContext = createContext();
+export const SelectedDayContext = createContext();
