+++ docs/src/pages/main_page/main.js
@@ -1,26 +1,29 @@
 
-import React, {Component} from "react";
-import {Calendar} from "../../components/calendar"
+import React from "react";
+import { useState } from "react";
+import Todo from './todo';
+import Calendar from './calendar';
+import { MonthContext } from "../../contexts/date_context";
+import { SelectedDayContext } from "../../contexts/date_context";
 
-export function renderMain(){
+export const Main = () => {
+    const [currentMonth, setCurrentMonth] = useState(new Date());
+    const [selectedDate, setSelectedDate] = useState(new Date());
+
+    // const value = useMemo(() => [counter, actions], [counter, actions]);
+
     return (
-        <div className="main">
-        <div className="main__contents">
-            <div id="main__contents--calendar">
-            <nav>
-                <div className="year-month"></div>
-                <div className="arrows">
-                <div id="prev"><i className="bi bi-chevron-left"></i></div>
-                <div id="next"><i className="bi bi-chevron-right"></i></div>
-                </div>
-            </nav>
-            <Calendar/>
+      <MonthContext.Provider value={[currentMonth, setCurrentMonth]}>
+        <SelectedDayContext.Provider value={[selectedDate, setSelectedDate]}>
+          <div className="main">
+            <div className="main__contents">
+              <Calendar />
+              <Todo />
             </div>
-
-
-        </div>
-        </div> 
+          </div>
+        </SelectedDayContext.Provider>
+      </MonthContext.Provider>
     );
-}
+};
 
-// export default renderMain;
+export default Main;
